migrated helper functions for probe IO
